refactor db names and file directory
