--- src/db/base_class.py
@@ -1,14 +1,0 @@
-from typing import Any
-
-from sqlalchemy.ext.declarative import declared_attr
-from sqlalchemy.orm import DeclarativeBase
-
-
-class Base(DeclarativeBase):
-    id: Any
-    __name__: str
-
-    # to generate tablename from classname
-    @declared_attr
-    def __tablename__(cls) -> str:
-        return cls.__name__.lower()
